feat: creates SDC version of existing icon.twig
Slightly changes the API to make it easier to use elsewhere without
needing to rely on the subdirectory structure of localgov_base:
icon_path in this version is a namespaced string representing the
*complete* path to the directory that actually contains the required
icons.

Existing template left in place and probably will be even if we replace
it in all localgov_base usage since we don't know what subthemes rely on
it.