Fix/1.x/705 lgd icon doesnt use theme variable (#706)
* feat: adds, uses new variable icon_fullpath

The new variable uses icon_path alone if it begins with '@' (i.e. if
it's Twig-namespaced), but will use the value of theme (defaulting to
'localgov_base') in other circumstances.

* feat: adds documentation for the 'theme' variable

Variable was present but unused, and undocumented outside of code
comments.

* feat: adds default value for undefined icon_path variable

This is the path that's used in localgov_base.

* feat: replaces variable assignmemt, use with in-place string derivation

Nothing actually changes but that we have three fewer lines of variable
assignment.